NCI-Agency/anet#1705: Drop SQLite support completely in favour of PostgreSQL
