+++ docs/anet.yml.production.template
@@ -118,7 +118,6 @@ logging:
           archivedLogFilenamePattern: ./logs/dbLog-%d.log.zip
           archivedFileCount: 2
           logFormat: '[%d{yyyy-MM-dd HH:mm:ss.SSS,UTC}] %p %c: %m%n'
-    "org.sqlite.JDBC" : INFO
     "io.dropwizard.assets.AssetsBundle" : INFO
     "io.dropwizard.assets.*" : INFO
     "waffle.servlet.NegotiateSecurityFilter" : INFO
