+++ insertBaseData-mssql.sql
@@ -886,7 +886,6 @@ INSERT INTO noteRelatedObjects (noteUuid, relatedObjectType, relatedObjectUuid)
 -- LEAVE THIS AS LAST STATEMENT
 -- Truncate all the dates (on reports etc.) to dates that could have been generated by
 -- Java (millisecond precision) rather than by the database itself (microsecond precision)
--- Note: keep SET on same line as UPDATE so mssql2sqlite.sh works correctly
 UPDATE reports SET
     createdAt=cast(createdAt as datetime2(3)),
     updatedAt=cast(updatedAt as datetime2(3)),
