+++ src/main/java/mil/dds/anet/database/ReportDao.java
@@ -74,8 +74,6 @@ public class ReportDao extends AnetBaseDao<Report> {
     switch (dbType) {
       case MSSQL:
         return "DATEPART(week, %s)";
-      case SQLITE:
-        return "strftime('%%%%W', substr(%s, 1, 10))";
       case POSTGRESQL:
         return "EXTRACT(WEEK FROM %s)";
       default:
@@ -567,8 +565,8 @@ public class ReportDao extends AnetBaseDao<Report> {
   }
 
   /**
-   * Helper method that builds and executes the daily rollup query Handles both MsSql and Sqlite
-   * Searching for just all reports and for reports in certain organizations.
+   * Helper method that builds and executes the daily rollup query. Searching for just all reports
+   * and for reports in certain organizations.
    * 
    * @param orgType the type of organization to be looking for
    * @param orgs the list of orgs for whose reports to find, null means all
@@ -588,13 +586,13 @@ public class ReportDao extends AnetBaseDao<Report> {
     sql.append("FROM reports WHERE ");
 
     // NOTE: more date-comparison work here that might be worth abstracting, but might not
-    if (getDbType() != DaoUtils.DbType.SQLITE) {
-      sql.append("\"releasedAt\" >= :startDate and \"releasedAt\" < :endDate "
-          + "AND \"engagementDate\" > :engagementDateStart ");
-    } else {
-      sql.append("\"releasedAt\"  >= DateTime(:startDate) AND \"releasedAt\" <= DateTime(:endDate) "
-          + "AND \"engagementDate\" > DateTime(:engagementDateStart) ");
-    }
+    // if (getDbType() != DaoUtils.DbType.SQLITE) {
+    sql.append("\"releasedAt\" >= :startDate and \"releasedAt\" < :endDate "
+        + "AND \"engagementDate\" > :engagementDateStart ");
+    // } else {
+    // sql.append("\"releasedAt\" >= DateTime(:startDate) AND \"releasedAt\" <= DateTime(:endDate) "
+    // + "AND \"engagementDate\" > DateTime(:engagementDateStart) ");
+    // }
     DaoUtils.addInstantAsLocalDateTime(sqlArgs, "startDate", start);
     DaoUtils.addInstantAsLocalDateTime(sqlArgs, "endDate", end);
     DaoUtils.addInstantAsLocalDateTime(sqlArgs, "engagementDateStart",
