+++ src/main/java/mil/dds/anet/search/Searcher.java
@@ -3,7 +3,6 @@ package mil.dds.anet.search;
 import com.google.inject.Injector;
 import mil.dds.anet.search.mssql.MssqlSearcher;
 import mil.dds.anet.search.pg.PostgresqlSearcher;
-import mil.dds.anet.search.sqlite.SqliteSearcher;
 import mil.dds.anet.utils.DaoUtils;
 
 public abstract class Searcher implements ISearcher {
@@ -18,8 +17,6 @@ public abstract class Searcher implements ISearcher {
     switch (dbType) {
       case MSSQL:
         return new MssqlSearcher(injector);
-      case SQLITE:
-        return new SqliteSearcher(injector);
       case POSTGRESQL:
         return new PostgresqlSearcher(injector);
       default:
